Split new Query into appropriate folders. Add to Controller.
