+++ src/Application/Users/Queries/GetNewUsers/GetNewUsersQuery.cs
@@ -1,22 +1,8 @@
 using AutoMapper.QueryableExtensions;
 using SSW.Rewards.Application.Common.Extensions;
+using SSW.Rewards.Shared.DTOs.Users;
 
 namespace SSW.Rewards.Application.Users.Queries.GetNewUsers;
-    
-// TODO: something went wrong with this query. It should just be returning users based on the filter, nothing to do with staff or achievements
-public class NewUserDto
-{
-    public int? UserId { get; set; }
-
-    public string? Name { get; set; }
-
-    public string? Email { get; set; }
-}
-
-public class NewUsersViewModel
-{
-    public IEnumerable<NewUserDto> NewUsers;
-}
 
 public class GetNewUsersQuery : IRequest<NewUsersViewModel>
 {
@@ -32,34 +18,33 @@ public class GetNewUsersHandler(
 {
     public async Task<NewUsersViewModel> Handle(GetNewUsersQuery request, CancellationToken cancellationToken)
     {
-        // find all the activated users with enough points in the (today/month/year/forever) leaderboard to claim the specific reward 
         var eligibleUsers = context.Users.Where(u => u.Activated == true);
 
         if (request.Filter == LeaderboardFilter.ThisYear)
         {
             eligibleUsers = eligibleUsers
-                .TagWith("PointsThisYear")
+                .TagWith("NewUsersThisYear")
                 .Where(u => u.CreatedUtc.Year == dateTime.Now.Year);
         }
         else if (request.Filter == LeaderboardFilter.ThisMonth)
         {
             eligibleUsers = eligibleUsers
-                .TagWith("PointsThisMonth")
+                .TagWith("NewUsersThisMonth")
                 .Where(u => u.CreatedUtc.Year == dateTime.Now.Year && u.CreatedUtc.Month == dateTime.Now.Month);
         }
         else if (request.Filter == LeaderboardFilter.Today)
         {
             eligibleUsers = eligibleUsers
-                .TagWith("PointsToday")
+                .TagWith("NewUsersToday")
                 .Where(u => u.CreatedUtc.Year == dateTime.Now.Year && u.CreatedUtc.Month == dateTime.Now.Month && u.CreatedUtc.Day == dateTime.Now.Day);
         }
         else if (request.Filter == LeaderboardFilter.ThisWeek)
         {
             var start = dateTime.Now.FirstDayOfWeek();
             var end = start.AddDays(7);
-            // TODO: Find a better way - EF Can't translate our extension method -- so writing the date range comparison directly in linq for now
+            
             eligibleUsers = eligibleUsers
-                .TagWith("PointsThisWeek")
+                .TagWith("NewUsersThisWeek")
                 .Where(u => start <= u.CreatedUtc && u.CreatedUtc <= end);
         }
         else if (request.Filter == LeaderboardFilter.Forever)
