+++ src/Common/DTOs/Users/NewUserDto.cs
@@ -1,0 +1,10 @@
+namespace SSW.Rewards.Shared.DTOs.Users;
+
+public class NewUserDto
+{
+    public int? UserId { get; set; }
+
+    public string? Name { get; set; }
+
+    public string? Email { get; set; }
+}
