+++ src/Common/DTOs/Users/NewUsersViewModel.cs
@@ -1,0 +1,6 @@
+namespace SSW.Rewards.Shared.DTOs.Users;
+
+public class NewUsersViewModel
+{
+    public IEnumerable<NewUserDto> NewUsers;
+}
